\name{describe_filter}
\alias{describe_filter}
\title{Gets the details of a filter configuration.}
\usage{
describe_filter(filters, afilter)
}
\arguments{
  \item{filters}{result of 'list_filters'}

  \item{afilter}{a specific filter name}
}
\value{
a character vector of named filter details
}
\description{
Gets the details of a filter configuration.
}
\seealso{
list_filters
}

