\name{get_filter_names}
\alias{get_filter_names}
\title{Extracts a list of filter names from the filter list as a result of 'list_filters'.}
\usage{
get_filter_names(filters)
}
\arguments{
  \item{filters}{result of 'list_filters'}
}
\value{
a vector of filter names
}
\description{
Extracts a list of filter names from the filter list as a
result of 'list_filters'.
}
\seealso{
list_filters
}

