\name{get_filter_values}
\alias{get_filter_values}
\title{Extracts a list of filter values from the filter list as a result of 'list_filters'.}
\usage{
get_filter_values(filters, afilter)
}
\arguments{
  \item{filters}{result of 'list_filters'}

  \item{afilter}{a specific filter name}
}
\value{
a vector of filter names
}
\description{
Extracts a list of filter values from the filter list as a
result of 'list_filters'.
}
\seealso{
list_filters
}

